+++ components/generator/jvm/imagevector/src/main/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/spec/LazyPropertySpec.kt
@@ -62,5 +62,6 @@ internal class LazyPropertySpec(private val config: ImageVectorSpecConfig) {
                 .endControlFlow()
                 .build(),
         )
+        addSuppressUnusedReceiverAnnotation(iconPackClassName)
     }
 }