+++ sdk/compose/icons/src/commonMain/kotlin/io/github/composegears/valkyrie/sdk/compose/icons/outlined/BatchProcessing.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.graphics.vector.path
 import androidx.compose.ui.unit.dp
 import io.github.composegears.valkyrie.sdk.compose.icons.ValkyrieIcons
 
+@Suppress("UnusedReceiverParameter")
 val ValkyrieIcons.Outlined.BatchProcessing: ImageVector
     get() {
         if (_BatchProcessing != null) {