+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/service/PersistentSettings.kt
@@ -51,6 +51,7 @@ class PersistentSettings : SimplePersistentStateComponent<ValkyrieState>(Valkyri
         var useExplicitMode: Boolean by property(false)
         var addTrailingComma: Boolean by property(false)
         var usePathDataString: Boolean by property(false)
+        var suppressUnusedReceiverWarning: Boolean by property(false)
 
         var showImageVectorPreview: Boolean by property(true)
         var showIconsInProjectView: Boolean by property(true)