+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/conversion/IconPackConversionViewModel.kt
@@ -376,6 +376,7 @@ class IconPackConversionViewModel(
             addTrailingComma = addTrailingComma,
             usePathDataString = usePathDataString,
             indentSize = indentSize,
+            suppressUnusedReceiverWarning = suppressUnusedReceiverWarning,
         )
     }
 }